add a stock component
